[JKImageFlowDataSource] Cleanup
This protocol was a mess.  The protocol methods were named after ImageBrowser rather than ImageFlow.  However, the methods implemented in the actual tests were based on ImageFlow and an number of work arounds were in place to make this work by force.  This change cleans it all up.

+++ iOS/JKImageFlowViewTest/JKImageFlowViewTest/ViewController.m
@@ -95,7 +95,7 @@
     return [images count];
 }
 
-- (id)imageFlow:(id)aFlow itemAtIndex:(int)index
+- (id)imageFlow:(id)aFlow itemAtIndex:(NSUInteger)index
 {
     return [images objectAtIndex:index];
 }
